requirements.txt
fastapi==0.115.12
passlib==1.7.4
pydantic==2.11.4
python-jose==3.4.0
SQLAlchemy==2.0.40
starlette~=0.46.2